AnimeGAN/requirements.txt
numpy
onnxruntime-gpu
torch 
torchvision
facenet_pytorch
gradio
opencv-python